# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASOTRACMET · Enturnamiento",
    "version": "1.0.0",
    "description": "API REST del sistema de enturnamiento (spec §8). Errores con código estable (`ErrorApi`); sesiones Bearer opacas y revocables; RBAC por recurso (`x-guard`); `x-reauth` exige re-autenticación reciente."
  },
  "servers": [
    {
      "url": "/"
    }
  ],
...
  "paths": {
...
    "/api/v1/documentos/{id}": {
      "patch": {
        "tags": ["HSEQ"],
        "summary": "Renovar o corregir documento",
        "operationId": "patchDocumentosId",
        "x-guard": "documentos U",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActualizarDocumento"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "VistaDocumento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VistaDocumento"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["HSEQ"],
        "summary": "Baja lógica",
        "operationId": "deleteDocumentosId",
        "x-guard": "documentos D",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "VistaDocumento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VistaDocumento"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/documentos/{id}/soporte": {
      "post": {
        "tags": ["HSEQ"],
        "summary": "Pedir dónde subir el soporte (URL prefirmada de S3 o esta API); PDF, JPG o PNG hasta 10 MB",
        "operationId": "postDocumentosIdSoporte",
        "x-guard": "documentos U",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SolicitarSoporte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "UrlSubida",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UrlSubida"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": ["HSEQ"],
        "summary": "Descargar el soporte con el rol del actor (302 a URL prefirmada en S3, bytes en local); auditado",
        "operationId": "getDocumentosIdSoporte",
        "x-guard": "documentos R (own)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "application/pdf o imagen (local), o 302 a la URL prefirmada (S3)",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "description": "application/pdf o imagen (local), o 302 a la URL prefirmada (S3)",
                  "x-tipo": "Soporte"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/soportes/{clave}": {
      "put": {
        "tags": ["HSEQ"],
        "summary": "Subida de bytes al almacén local (con S3 el navegador sube directo al bucket)",
        "operationId": "putSoportesClave",
        "x-guard": "documentos U",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "clave",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "SoporteGuardado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SoporteGuardado"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/documentos/{id}/soporte/confirmar": {
      "post": {
        "tags": ["HSEQ"],
        "summary": "Confirmar la subida: verifica el objeto y guarda la referencia en `archivoUrl` (auditado)",
        "operationId": "postDocumentosIdSoporteConfirmar",
        "x-guard": "documentos U",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmarSoporte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "VistaDocumento",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VistaDocumento"
                }
              }
            }
          },
          "default": {
            "description": "Error con código estable (spec §8.6): 400 VALIDATION_ERROR, 401, 403, 404, 409 o 422 según el código",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorApi"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer"
      }
    },
    "schemas": {
      "ErrorApi": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "UNAUTHORIZED",
              "CHALLENGE_INVALIDO",
              "CODIGO_INVALIDO",
              "FORBIDDEN",
              "FORBIDDEN_OWN_SCOPE",
              "REAUTH_REQUERIDA",
              "NOT_FOUND",
              "TOTP_NO_ENROLADO",
              "COLA_LOCKED",
              "EMAIL_EN_USO",
              "PLACA_EN_USO",
              "DOCUMENTO_EN_USO",
              "CATALOGO_EN_USO",
              "TARIFA_DUPLICADA",
              "VEHICULO_CON_HISTORIA",
              "TR_DUPLICADO",
              "OFERTA_NO_ABIERTA",
              "REQUERIMIENTO_CERRADO",
              "REQUERIMIENTO_SIN_CUPOS",
              "TR_NO_CANCELABLE",
              "VERSION_CONFLICT",
              "VIAJE_YA_EXISTE",
              "VIAJE_LIQUIDADO",
              "VIAJE_ANULADO",
              "RECAUDO_CERRADO",
              "COLA_VACIA",
              "VEHICULO_NO_HABILITADO",
              "VEHICULO_NO_ACTIVO",
              "DOCUMENTO_VENCIDO",
              "BLOQUEO_TEMPORAL",
              "OFERTA_ABIERTA_PREVIA",
              "TR_ACTIVO",
              "PENALIZACION_PENDIENTE",
              "OFERTA_EXPIRADA",
              "MOTIVO_REQUERIDO",
              "TR_NO_VIAJABLE",
              "VIAJE_NO_LIQUIDABLE",
              "PAGO_INVALIDO",
              "RATE_LIMITED",
              "DEMASIADOS_INTENTOS",
              "INVARIANTE_COLA",
              "INTERNAL"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {}
          }
        },
        "required": ["code", "message"],
        "additionalProperties": false
      },
...
      "VistaDocumento": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "sujetoTipo": {
            "type": "string",
            "enum": ["vehiculo", "conductor"]
          },
          "sujetoId": {
            "type": "string"
          },
          "tipoId": {
            "type": "string"
          },
          "numero": {
            "type": ["string", "null"]
          },
          "emitidoEn": {
            "type": ["string", "null"]
          },
          "venceEn": {
            "type": ["string", "null"]
          },
          "archivoUrl": {
            "type": ["string", "null"]
          },
          "eliminadoEn": {
            "type": ["string", "null"]
          },
          "tipo": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "id": {
                    "type": "string"
                  },
                  "codigo": {
                    "type": "string"
                  },
                  "nombre": {
                    "type": "string"
                  },
                  "aplicaA": {
                    "type": "string",
                    "enum": ["vehiculo", "conductor"]
                  },
                  "bloqueante": {
                    "type": "boolean"
                  },
                  "diasAlerta": {
                    "type": "number"
                  }
                },
                "required": ["id", "codigo", "nombre", "aplicaA", "bloqueante", "diasAlerta"],
                "additionalProperties": false
              },
              {
                "type": "null"
              }
            ]
          },
          "estado": {
            "type": "string",
            "enum": ["vigente", "por_vencer", "vencido"]
          },
          "diasParaVencer": {
            "type": ["number", "null"]
          }
        },
        "required": [
          "id",
          "sujetoTipo",
          "sujetoId",
          "tipoId",
          "numero",
          "emitidoEn",
          "venceEn",
          "archivoUrl",
          "eliminadoEn",
          "tipo",
          "estado",
          "diasParaVencer"
        ],
        "additionalProperties": false
      },
...
      "ActualizarDocumento": {
        "type": "object",
        "properties": {
          "numero": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 60
              },
              {
                "type": "null"
              }
            ]
          },
          "emitidoEn": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "venceEn": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^\\d{4}-\\d{2}-\\d{2}$"
              },
              {
                "type": "null"
              }
            ]
          },
          "archivoUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500,
                "format": "uri"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "UrlSubida": {
        "type": "object",
        "properties": {
          "clave": {
            "type": "string"
          },
          "metodo": {
            "type": "string",
            "const": "PUT"
          },
          "url": {
            "type": "string"
          },
          "cabeceras": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "string"
            }
          },
          "expiraEn": {
            "type": "string"
          }
        },
        "required": ["clave", "metodo", "url", "cabeceras", "expiraEn"],
        "additionalProperties": false
      },
      "SolicitarSoporte": {
        "type": "object",
        "properties": {
          "nombre": {
            "type": "string",
            "minLength": 1,
            "maxLength": 160
          },
          "tipo": {
            "type": "string",
            "enum": ["application/pdf", "image/jpeg", "image/png"]
          },
          "tamano": {
            "type": "integer",
            "exclusiveMinimum": 0,
            "maximum": 10485760
          }
        },
        "required": ["nombre", "tipo", "tamano"]
      },
      "SoporteGuardado": {
        "type": "object",
        "properties": {
          "clave": {
            "type": "string"
          },
          "tamano": {
            "type": "number"
          }
        },
        "required": ["clave", "tamano"],
        "additionalProperties": false
      },
      "ConfirmarSoporte": {
        "type": "object",
        "properties": {
          "clave": {
            "type": "string",
            "pattern": "^documentos\\/[A-Za-z0-9_-]{1,64}\\/[A-Za-z0-9_-]{1,64}\\.(pdf|jpg|png)$"
          }
        },
        "required": ["clave"]
      },
...
    }
  }
}